% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ammit_spe_dummy}
\alias{ammit_spe_dummy}
\title{AMMIT dummy SPE data}
\format{
\subsection{\code{ammit_spe_dummy}}{

A SpatialExperiment object with 7 rows and 9,906 columns:
}
}
\usage{
ammit_spe_dummy
}
\description{
A subset of a HALO skin cancer dataset. There are 7 protein markers (names censored) across 10,000 cells (filtered to 9,906 DAPI-positive cells).
There are two anaylsis regions, Tumour and Stroma; and two distinct tissue fragments.
}
\keyword{datasets}
