% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ammit_thresholds.R
\name{get_ammit_thresholds}
\alias{get_ammit_thresholds}
\title{Derive the AMMIT threshold from unmixed marker intensity data and a given strictness parameter.}
\usage{
get_ammit_thresholds(object, markers, strictness, return = "object", ...)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present, or a list of SpatialExperiment objects.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object for which to derive thresholds, or "all" (default) to do so for all markers.}

\item{strictness}{Numeric; one or multiple strictness values matching to each of the marker(s) specified.}

\item{return}{Character; either 'thresholds' to return a named vector of the derived thresholds for each marker, or 'object' (default) to return the SpatialExperiment object with this information embedded in the rowData.}

\item{...}{Arguments to pass to infer_strictness_split (subset) for a list input}
}
\value{
A single, or list of: Numeric vector(s) of derived thresholds (if 'thresholds' specified for return), or SpatialExperiment object(s) with this information embedded in rowData (if 'object' specified).
}
\description{
Derive the AMMIT threshold from unmixed marker intensity data and a given strictness parameter.
}
\examples{
data("ammit_spe_dummy")
ammit_spe_dummy <- infer_manual_thresholds(ammit_spe_dummy, markers = "M7")
ammit_spe_dummy <- unmix_intensities(ammit_spe_dummy, markers="M7", k=3)
strictness <- infer_strictness(ammit_spe_dummy, markers = "M7", return = "strictness")
get_ammit_thresholds(ammit_spe_dummy, markers = "M7", strictness = strictness, return = "thresholds")
}
