% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ammit_thresholds.R
\name{get_ammit_thresholds_split}
\alias{get_ammit_thresholds_split}
\title{Derive the AMMIT threshold from unmixed marker intensity data and a given strictness parameter.}
\usage{
get_ammit_thresholds_split(
  object,
  markers,
  strictness,
  return = "object",
  subset = NULL
)
}
\arguments{
\item{object}{A list of SpatialExperiment objects, each with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object for which to derive thresholds, or "all" (default) to do so for all markers.}

\item{strictness}{Numeric; one or multiple strictness values matching to each of the marker(s) specified.}

\item{return}{Character; either 'thresholds' to return a list of named vectors of the derived thresholds for each marker, or 'object' (default) to return the list of SpatialExperiment objects with this information embedded in the rowData.}

\item{subset}{Character; Which of the split subsets for which to derive thresholds. If left NULL (default), all subsets will have strictness inferred.}
}
\value{
A list of: Numeric vectors of derived thresholds (if 'thresholds' specified for return), or SpatialExperiment objects with this information embedded in rowData (if 'object' specified).
}
\description{
Derive the AMMIT threshold from unmixed marker intensity data and a given strictness parameter.
}
\examples{
data("ammit_spe_dummy")
spe_list <- split_spe(ammit_spe_dummy)
spe_list <- infer_manual_thresholds(spe_list, markers = "M7")
spe_list <- unmix_intensities(spe_list, markers="M7", k=3)
strictness <- infer_strictness(spe_list, markers = "M7", return = "strictness")
get_ammit_thresholds(spe_list, markers = "M7", strictness = strictness, return = "thresholds")
}
