% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_manual_thresholds.R
\name{infer_manual_thresholds}
\alias{infer_manual_thresholds}
\title{Infer manual thresholds for marker intensities from pre-existing classifications}
\usage{
infer_manual_thresholds(spe, markers)
}
\arguments{
\item{spe}{A SpatialExperiment object with assays 'data' and 'reference' present.}

\item{markers}{A character vector of markers matching, at least partially, to the rownames of your SpatialExperiment object.}
}
\value{
The input SpatialExperiment object, with manual thresholds for each of the elements in markers added into the rowData under manual_thresholds
}
\description{
Infer manual thresholds for marker intensities from pre-existing classifications
}
\examples{
data("ammit_spe_dummy")
markers <- c("M1", "M2", "M3", "M4", "M5", "M6", "M7")
ammit_spe_dummy <- infer_manual_thresholds(ammit_spe_dummy, markers=markers)
SummarizedExperiment::rowData(ammit_spe_dummy)
}
