% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_manual_thresholds.R
\name{infer_manual_thresholds_split}
\alias{infer_manual_thresholds_split}
\title{Infer manual thresholds for marker intensities from pre-existing classifications on a list of SpatialExperiment objects}
\usage{
infer_manual_thresholds_split(object, markers, subset = NULL)
}
\arguments{
\item{object}{A list of SpatialExperiment objects, each with assays 'data' and 'reference' present.}

\item{markers}{A character vector of markers matching, at least partially, to the rownames of your SpatialExperiment objects, or "all" (default) for all markers.}

\item{subset}{A subset of the names of the list of objects for which you wish to infer manual thresholds.}
}
\value{
The input list of SpatialExperiment objects, each with manual thresholds for all markers specified inserted in the rowData under manual_thresholds.
}
\description{
Infer manual thresholds for marker intensities from pre-existing classifications on a list of SpatialExperiment objects
}
\examples{
data("ammit_spe_dummy")
spe_list <- split_spe(ammit_spe_dummy, split.by="Analysis.Region")
markers <- c("M1", "M2", "M3", "M4", "M5", "M6", "M7")
spe_list <- infer_manual_thresholds(spe_list, markers=markers)
lapply(spe_list, SummarizedExperiment::rowData)
}
