% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_strictness.R
\name{infer_strictness}
\alias{infer_strictness}
\title{Infer the AMMIT strictness parameter required to derive manual threshold values from unmixed marker intensity data}
\usage{
infer_strictness(object, markers = "all", return = "strictness", ...)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present, or a list of SpatialExperiment objects.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object for which to infer strictness, or "all" (default) to do so for all markers.}

\item{return}{Character; either 'strictness' (default) to return a named vector of the inferred strictness for each marker, or 'object' to return the SpatialExperiment object with this information embedded in the rowData.}

\item{...}{Arguments to pass to infer_strictness_split (subset) for a list input}
}
\value{
A single, or list of: Numeric vector(s) of inferred strictness values (if 'strictness' specified for return), or SpatialExperiment object(s) with this information embedded in rowData (if 'object' specified).
}
\description{
Infer the AMMIT strictness parameter required to derive manual threshold values from unmixed marker intensity data
}
\examples{
data("ammit_spe_dummy")
ammit_spe_dummy <- infer_manual_thresholds(ammit_spe_dummy, markers = "all")
ammit_spe_dummy <- unmix_intensities(ammit_spe_dummy, markers="all", k=3)
infer_strictness(ammit_spe_dummy, markers = "all", return = "strictness")
}
