% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensities.R
\name{plot_intensities_single}
\alias{plot_intensities_single}
\title{Plot a single marker intensity from an AMMIT SpatialExperiment Object}
\usage{
plot_intensities_single(
  object,
  marker,
  thresholds = "none",
  transformation = "asinh",
  censor_zeroes = TRUE
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present.}

\item{marker}{Character; A single marker present in the rownames of the SpatialExperiment Object, to be plotted.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{transformation}{Character; Passed to transform_intensities. Either "none" to use raw values, or one of "asinh" or "sqrt" to apply the respective transformation to the data. Defaults to "asinh".}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}
}
\value{
A single ggplot density plot of the intensities for the marker selected.
}
\description{
Plot a single marker intensity from an AMMIT SpatialExperiment Object
}
\examples{
#' data("ammit_spe_dummy")
plot_intensities_spe(ammit_spe_dummy,
                 marker = "M7",
                 thresholds = "none",
                 transformation = "asinh",
                 censor_zeroes = TRUE)
}
