% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensities.R
\name{plot_intensities_spe}
\alias{plot_intensities_spe}
\title{Plot the marker intensities of an AMMIT SpatialExperiment Object}
\usage{
plot_intensities_spe(
  object,
  markers = "all",
  thresholds = "none",
  transformation = "asinh",
  censor_zeroes = TRUE
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to plot, or "all" (default) to plot all markers.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{transformation}{Character; Passed to transform_intensities. Either "none" to use raw values, or one of "asinh" or "sqrt" to apply the respective transformation to the data. Defaults to "asinh".}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}
}
\value{
A ggplot density plot for each of the markers selected.
}
\description{
Plot the marker intensities of an AMMIT SpatialExperiment Object
}
\examples{
data("ammit_spe_dummy")
plot_intensities_spe(ammit_spe_dummy,
                 markers = "M7",
                 thresholds = "none",
                 transformation = "asinh",
                 censor_zeroes = TRUE)
}
