% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensities.R
\name{plot_intensities_split}
\alias{plot_intensities_split}
\title{Plot the marker intensities for a list of SpatialExperiment Objects}
\usage{
plot_intensities_split(
  object,
  markers = "all",
  thresholds = "none",
  transformation = "asinh",
  censor_zeroes = TRUE,
  subset = NULL
)
}
\arguments{
\item{object}{A list of SpatialExperiment objects, each with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to plot, or "all" (default) to plot all markers.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{transformation}{Character; Passed to transform_intensities. Either "none" to use raw values, or one of "asinh" or "sqrt" to apply the respective transformation to the data. Defaults to "asinh".}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}

\item{subset}{Character; Which of the split subsets to plot. If left NULL (default), all subsets will be plotted.}
}
\value{
A list of patchworked ggplot density plots with each of the markers selected, for each subset of the list.
}
\description{
Plot the marker intensities for a list of SpatialExperiment Objects
}
\examples{
data("ammit_spe_dummy")
spe_list <- split_spe(ammit_spe_dummy, split.by="Analysis.Region")
plot_intensities_split(spe_list,
                 markers = "M7",
                 thresholds = "none",
                 transformation = "asinh",
                 censor_zeroes = TRUE,
                 subset = "Tumour")
}
