% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture.R
\name{plot_mixture}
\alias{plot_mixture}
\title{Plot the unmixed distributions of marker intensities of an AMMIT SpatialExperiment Object}
\usage{
plot_mixture(
  object,
  markers = "all",
  thresholds = "none",
  combine_negatives = TRUE,
  censor_zeroes = TRUE,
  ...
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present, or a list of SpatialExperiment objects.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to plot, or "all" (default) to plot all markers.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{combine_negatives}{Logical; Whether to combine the negative distributions, if there are multiple, to plot them as one curve (on which the ammit threshold would have been calculated)}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}

\item{...}{Arguments to be passed to plot_mixture_split in the case of a list of SpatialExperiment objects.}
}
\value{
A ggplot density plot for each of the markers selected
}
\description{
Plot the unmixed distributions of marker intensities of an AMMIT SpatialExperiment Object
}
\examples{
data("ammit_spe_dummy")
ammit_spe_dummy <- infer_manual_thresholds(ammit_spe_dummy, markers = "M7")
ammit_spe_dummy <- unmix_intensities(ammit_spe_dummy, markers="M7", k=2)
ammit_spe_dummy <- get_ammit_thresholds(ammit_spe_dummy,
                                        markers = "M7",
                                        strictness = 0.1,
                                        return = "object")
plot_mixture(ammit_spe_dummy,
             markers = "M7",
             thresholds = "both",
             combine_negative = FALSE,
             censor_zeroes = TRUE)
}
