% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture.R
\name{plot_mixture_single}
\alias{plot_mixture_single}
\title{Plot the AMMIT-unmixed distributions of a marker's intensity values}
\usage{
plot_mixture_single(
  object,
  marker,
  thresholds = "none",
  combine_negatives = TRUE,
  censor_zeroes = TRUE
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present.}

\item{marker}{Character; A marker, found in the SpatialExperiment object, to plot.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{combine_negatives}{Logical; Whether to combine the negative distributions, if there are multiple, to plot them as one curve (on which the ammit threshold would have been calculated)}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}
}
\value{
A single ggplot histogram plot of the intensities for the marker selected, with the unmixed distributions plotted on top.
}
\description{
Plot the AMMIT-unmixed distributions of a marker's intensity values
}
\examples{
data("ammit_spe_dummy")
ammit_spe_dummy <- infer_manual_thresholds(ammit_spe_dummy, markers = "M7")
ammit_spe_dummy <- unmix_intensities(ammit_spe_dummy, markers="M7", k=2)
ammit_spe_dummy <- get_ammit_thresholds(ammit_spe_dummy,
                                        markers = "M7",
                                        strictness = 0.1,
                                        return = "object")
plot_mixture_single(ammit_spe_dummy,
                    marker = "M7",
                    thresholds = "both",
                    combine_negative = FALSE,
                    censor_zeroes = TRUE)
}
