% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture.R
\name{plot_mixture_spe}
\alias{plot_mixture_spe}
\title{Plot the unmixed distributions of marker intensities of an AMMIT SpatialExperiment Object}
\usage{
plot_mixture_spe(
  object,
  markers = "all",
  thresholds = "none",
  combine_negatives = TRUE,
  censor_zeroes = TRUE
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to plot, or "all" (default) to plot all markers.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{combine_negatives}{Logical; Whether to combine the negative distributions, if there are multiple, to plot them as one curve (on which the ammit threshold would have been calculated)}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}
}
\value{
A ggplot density plot for each of the markers selected.
}
\description{
Plot the unmixed distributions of marker intensities of an AMMIT SpatialExperiment Object
}
