% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture.R
\name{plot_mixture_split}
\alias{plot_mixture_split}
\title{Plot the unmixed distributions of marker intensities for a list of SpatialExperiment Objects}
\usage{
plot_mixture_split(
  object,
  markers = "all",
  thresholds = "none",
  combine_negatives = TRUE,
  censor_zeroes = TRUE,
  subset = NULL
)
}
\arguments{
\item{object}{A list of SpatialExperiment objects, each with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to plot, or "all" (default) to plot all markers.}

\item{thresholds}{Character; "none" to plot no threshold values, "manual" to plot manual thresholds (see infer_manual_thresholds), "ammit" to plot ammit thresholds, or "both" for both. Defaults to "none".}

\item{combine_negatives}{Logical; Whether to combine the negative distributions, if there are multiple, to plot them as one curve (on which the ammit threshold would have been calculated)}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before plotting. Recommended; defaults to TRUE.}

\item{subset}{Character; Which of the split subsets to plot. If left NULL (default), all subsets will be plotted.}
}
\value{
A list of patchworked ggplot density plots with each of the markers selected, for each subset of the list.
}
\description{
Plot the unmixed distributions of marker intensities for a list of SpatialExperiment Objects
}
