% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spe.R
\name{split_spe}
\alias{split_spe}
\title{Split a SpatialExperiment object based on a variable}
\usage{
split_spe(spe, split.by = "Analysis.Region")
}
\arguments{
\item{spe}{A SpatialExperiment object to be split.}

\item{split.by}{Character; A column name present in the colData of spe, to split upon. Defaults to "Analysis.Region".}
}
\value{
A named list of SpatialExperiment objects, where each object is the subset of the original for one value of the splitting variable.
}
\description{
Split a SpatialExperiment object based on a variable
}
\examples{
data("ammit_spe_dummy")
split_spe(ammit_spe_dummy, split.by="Analysis.Region")
}
