% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_intensities.R
\name{transform_intensities}
\alias{transform_intensities}
\title{Transform marker intensities for unmixing or plotting}
\usage{
transform_intensities(intensities, method)
}
\arguments{
\item{intensities}{Numeric; a vector of marker intensities (or an intensity threshold)}

\item{method}{Character; one of: 'none' for no transformation, 'asinh' for arcsinh transformation, or 'sqrt' for square-root transformation.}
}
\value{
A transformed numeric vector.
}
\description{
Transform marker intensities for unmixing or plotting
}
\examples{
intensities <- c(1,3,3,4,5,3,3)
transform_intensities(intensities, method = "asinh")
}
