% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_intensities.R
\name{unmix_intensities}
\alias{unmix_intensities}
\title{Unmix marker intensity data from an AMMIT SpatialExperiment object into multiple skew-normal distributions}
\usage{
unmix_intensities(
  object,
  markers,
  k,
  downsample = 15000,
  transformation = "asinh",
  censor_zeroes = TRUE,
  ...
)
}
\arguments{
\item{object}{A SpatialExperiment object with assay 'data' present, or a list of SpatialExperiment objects.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to unmix, or "all" (default) to unmix all markers.}

\item{k}{Numeric; How many mixtures to derive from the intensity data. Will be recycled to match the number of markers. Markers can have different k values!}

\item{downsample}{Numeric; If provided, how many cells to downsample intensity data to before unmixing. Default value is 15000. If NULL or non-numeric, no downsampling will be done.}

\item{transformation}{Character; Passed to transform_intensities. Either "none" to use raw values, or one of "asinh" or "sqrt" to apply the respective transformation to the data. Defaults to "asinh".}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before unmixing. Recommended; defaults to TRUE.}

\item{...}{Arguments to pass to unmix_intensites_split (subset) for a list input; or to smsn.mix, such as nu and iter.max.}
}
\value{
The input object(s), with model parameters for the unmixed distributions for each marker located in their respective rowData.
}
\description{
Unmix marker intensity data from an AMMIT SpatialExperiment object into multiple skew-normal distributions
}
\examples{
data("ammit_spe_dummy")
ammit_spe_dummy <- unmix_intensities(ammit_spe_dummy, markers="M7", k=2)
SummarizedExperiment::rowData(ammit_spe_dummy)["M7", ]
}
