% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_intensities.R
\name{unmix_intensities_split}
\alias{unmix_intensities_split}
\title{Unmix marker intensity data from a list of SpatialExperiment objects into multiple skew-normal distributions}
\usage{
unmix_intensities_split(
  object,
  markers,
  k,
  downsample = 15000,
  transformation = "asinh",
  censor_zeroes = TRUE,
  subset = NULL,
  ...
)
}
\arguments{
\item{object}{A list of SpatialExperiment objects, each with assay 'data' present.}

\item{markers}{Character; one or multiple markers found in the SpatialExperiment object to unmix, or "all" (default) to unmix all markers.}

\item{k}{Numeric; How many mixtures to derive from the intensity data. Will be recycled to match the number of markers. Markers can have different k values!}

\item{downsample}{Numeric; If provided, how many cells to downsample intensity data to before unmixing. Default value is 15000. If NULL or non-numeric, no downsampling will be done.}

\item{transformation}{Character; Passed to transform_intensities. Either "none" to use raw values, or one of "asinh" or "sqrt" to apply the respective transformation to the data. Defaults to "asinh".}

\item{censor_zeroes}{Logical; Whether to remove zeroes from intensity data before unmixing. Recommended; defaults to TRUE.}

\item{subset}{Character; Which of the split subsets to unmix. If left NULL (default), all subsets will be plotted.}

\item{...}{Arguments to pass to smnsn.mix; such as nu and iter.max.}
}
\value{
The input objects, with model parameters for the unmixed distributions for each marker located in their respective rowData.
}
\description{
Unmix marker intensity data from a list of SpatialExperiment objects into multiple skew-normal distributions
}
\examples{
data("ammit_spe_dummy")
spe_list <- split_spe(ammit_spe_dummy, split.by="Analysis.Region")
spe_list <- unmix_intensities(spe_list, markers="M7", k=2, subset="Tumour")
SummarizedExperiment::rowData(spe_list[["Tumour"]])["M7", ]
}
