% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ammit_spe_HALO.R
\name{build_ammit_spe_HALO}
\alias{build_ammit_spe_HALO}
\title{Title}
\usage{
build_ammit_spe_HALO(
  HALO_filepath,
  markers,
  reference,
  locations = NULL,
  filter_dapi = TRUE
)
}
\arguments{
\item{HALO_filepath}{Character; path to your HALO output file (.csv or .tsv), generally named 'Total_Object_Results.csv'}

\item{markers}{Character; a vector of the markers (i.e., proteins being measured) in the HALO output}

\item{reference}{Logical; whether the input file has manual thresholds computed, i.e. whether there are columns named "marker.Positive.Classification" or similar.}

\item{locations}{Character; a vector of the locations associated with each marker (e.g. Cytoplasm, Nucleus) for selecting the desired intensity data, if there are multiple intensity columns per marker.}

\item{filter_dapi}{Logical (Default TRUE); whether to filter the input to only DAPI-positive cells.}
}
\value{
A SpatialExperiment object with assay 'data' containing the intensities per cell for each marker, spatial coordinates, metadata (in colData), and positive classification (if provided) in the assay 'reference'.
}
\description{
Title
}
